[rename]: login→auth
