+++ assets/sass/app.scss
@@ -8,4 +8,4 @@
 @import "elements/buttons";
 @import "elements/transitions";
 
-@import "login/login.scss";
+@import "auth/auth.scss";
