+++ assets/sass/auth/auth.scss
@@ -1,4 +1,4 @@
-.login-container {
+.auth-container {
   background-color: #fff;
   background-image: url("~assets/img/ArchiWalk_background.jpeg");
   background-position: center center;
@@ -7,11 +7,11 @@
   background-size: 100% 100%;
 }
 
-.login-form {
+.auth-form {
   background-color: rgba(#fff, 0.8);
 }
 
-.login-form-title {
+.auth-form-title {
   padding: 1rem;
 }
