+++ components/auth/OtherPageLink.vue
